% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\name{corrplot}
\alias{corrplot}
\title{select independent variables and draw correlation plot}
\usage{
corrplot(dataset, depvar, n)
}
\arguments{
\item{dataset}{The dataset with many variables.}

\item{depvar}{The dependent variable.}

\item{n}{The number of independent variables.}
}
\description{
In a dataset with many variables, select several variables with the highest correlation with one specific variable, which can be the dependent variable of the study.
             Draw the correlation plot of these selected independent variables.
}
\examples{
corrplot(house_data, "SalePrice", 10)
}
