% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_NA.R
\name{drop_NA}
\alias{drop_NA}
\title{drop_NA}
\usage{
drop_NA(df, threshold = 0.1, null = NA)
}
\arguments{
\item{df}{The manipulated data frame.}

\item{threshold}{A threshold value that tolerates null values.
threshold=0.1 by default.}

\item{null}{The format of invalid values, such as NA, NULL, "...", etc.}
}
\value{
Return the processed data frame.
}
\description{
Drop the null values in a data frame.
}
\examples{
world_data=drop_NA(world_data_raw,null="..")
}
