% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fred.R
\name{get_fred}
\alias{get_fred}
\title{get_fred}
\usage{
get_fred(series_id, start_date, end_date)
}
\arguments{
\item{series_id}{The series id in FRED.}

\item{start_date}{The start date of this time series.}

\item{end_date}{The end date of this time series.}
}
\value{
Return a dataframe containing the time series as well as corresponding time.
}
\description{
Get time series from FRED, using API
}
\examples{
maize_price_df = get_fred_series("PMAIZMTUSDM","2011-10-01","2020-10-01")

}
