% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{imputation}
\usage{
imputation(df, col, method = "k-nearest")
}
\arguments{
\item{df}{The manipulated data frame.}

\item{col}{The column with missing data needing imputing.}

\item{method}{The imputation method, "k-nearest", "mean", "median".
method="k-nearest" by default.}
}
\value{
Return the processed data frame.
}
\description{
Several methods to impute the missing data
}
\examples{
world_data=drop_NA(world_data_raw,null="..")

world=imputation(world_data,col=c(4,7))
world=imputation(world_data,col=c(2:7),method = "mean")
world=imputation(world_data,col=c(4:5),method = "median")
}
