% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rand.R
\name{plot_rand}
\alias{plot_rand}
\title{A random series plot}
\usage{
plot_rand()
}
\description{
Generate two series of normally distributed random number, each with ten numbers, and plot.
}
\examples{
plot_rand()
}
